// src/cyfs-stack-test-typescript/cyfs_node/cyfs_node.d.ts
    interface Uint8Array {
        offset(offset: number): Uint8Array;
        offsetUint16Array(offset: number): Uint16Array;
        offsetUint32Array(offset: number): Uint32Array;
        offsetView(offset: number): DataView;
        toHex(): string;
        fromHex(hex: string): BuckyResult<Uint8Array>;
    }
// ...
export type Option<T> = (SomeOption<T> | NoneOption) & BaseOption<T>;
// ...
export interface RawEncode {
        raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        encode_to_buf?: (purpose?: RawEncodePurpose) => BuckyResult<Uint8Array>;
        raw_hash_encode?: () => BuckyResult<HashValue>;
}
// ...
export interface Compareable<T> {
        hashCode(): symbol;
        equals(other: T): boolean;
}
// ...
export type BuckyResult<T> = Result<T, BuckyError>;
// ...
export class HashValue implements RawEncode {
    m_buf: Uint8Array;
    constructor(buf: Uint8Array);
    as_slice(): Uint8Array;
    length(): number;
    static default(): HashValue;
    static copy_from_slice(buf: Uint8Array): HashValue;
    static hash_data(data: Uint8Array): HashValue;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    to_base_58(): string;
    to_hex_string(): string;
    static from_hex_string(s: string): BuckyResult<HashValue>;
}
// ...
export class ChunkDescContent extends DescContent {
    type_info(): DescTypeInfo;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class ChunkBodyContent extends BodyContent {
    type_info(): DescTypeInfo;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class ChunkId implements RawEncode {
    constructor(buf: Uint8Array);
    calculate_id(): ObjectId;
    obj_type(): number;
    obj_type_code(): number;
    as_slice(): Uint8Array;
    length(): number;
    toString(): string;
    to_string(): string;
    to_base_58(): string;
    static from_base_58(s: string): BuckyResult<ChunkId>;
    static default(): ChunkId;
    static copy_from_slice(buf: Uint8Array): ChunkId;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    encode_to_buf(): BuckyResult<Uint8Array>;
    desc(): ObjectDesc;
    body(): Option<ObjectMutBody<ChunkDescContent, ChunkBodyContent>>;
    signs(): ObjectSigns;
    nonce(): Option<JSBI>;
    static calculate(data: Uint8Array): BuckyResult<ChunkId>;
}
// ...
export class FileDescContent extends DescContent {
    len: JSBI;
    hash: HashValue;
    constructor(len: JSBI, hash: HashValue);
    type_info(): DescTypeInfo;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class FileDescContentDecoder extends DescContentDecoder<FileDescContent> {
    type_info(): DescTypeInfo;
    raw_decode(buf: Uint8Array): BuckyResult<[FileDescContent, Uint8Array]>;
}
export enum ChunkBundleHashMethod {
    Serial = "Serial"
}
export class ChunkBundle {
    chunk_list: ChunkId[];
    hash_method: ChunkBundleHashMethod;
    constructor(chunk_list: ChunkId[], hash_method: ChunkBundleHashMethod);
    len(): JSBI;
    calc_hash_value(): HashValue;
    calc_serial_hash_value(): HashValue;
}
export class ChunkList {
    chunk_in_list?: ChunkId[] | undefined;
    file_id?: FileId | undefined;
    chunk_in_bundle?: ChunkBundle | undefined;
    constructor(chunk_in_list?: ChunkId[] | undefined, file_id?: FileId | undefined, chunk_in_bundle?: ChunkBundle | undefined);
    inner_chunk_list(): ChunkId[] | undefined;
    match<T>(visitor: {
        ChunkInList?: (list: ChunkId[]) => T;
        FileId?: (id: FileId) => T;
        ChunkInBundle?: (bundle: ChunkBundle) => T;
    }): T | undefined;
}
export class FileBodyContent extends ProtobufBodyContent {
    chunk_list: ChunkList;
    constructor(chunk_list: ChunkList);
    inner_chunk_list(): ChunkId[] | undefined;
    try_to_proto(): BuckyResult<protos.FileBodyContent>;
}
// ...
export class FileId extends NamedObjectId<FileDescContent, FileBodyContent> {
    constructor(id: ObjectId);
    static default(): FileId;
    static from_base_58(s: string): BuckyResult<FileId>;
    static try_from_object_id(id: ObjectId): BuckyResult<FileId>;
}
export class FileIdDecoder extends NamedObjectIdDecoder<FileDescContent, FileBodyContent> {
    constructor();
}
export class File extends NamedObject<FileDescContent, FileBodyContent> {
    static create(owner: ObjectId, len: JSBI, hash: HashValue, chunk_list: ChunkList, build?: (builder: FileBuilder) => void): File;
    file_id(): FileId;
}
export class FileDecoder extends NamedObjectDecoder<FileDescContent, FileBodyContent, File> {
    constructor();
}
// ...
export abstract class ObjectDesc {
        constructor(obj_type: number);
        obj_type(): number;
        obj_type_code(): ObjectTypeCode;
        is_standard_object(): boolean;
        is_core_object(): boolean;
        is_dec_app_object(): boolean;
        abstract calculate_id(): ObjectId;
        abstract dec_id(): Option<ObjectId>;
        abstract ref_objs(): Option<Vec<ObjectLink>>;
        abstract prev(): Option<ObjectId>;
        abstract create_timestamp(): Option<HashValue>;
        abstract create_time(): JSBI;
        abstract expired_time(): Option<JSBI>;
        abstract owner(): Option<ObjectId> | undefined;
}
// ...
export class ObjectMutBody<DC extends DescContent, BC extends BodyContent> implements RawEncode {
        toString(): string;
        constructor(obj_type: number, prev_version: Option<HashValue>, update_time: JSBI, content: BC, user_data: Option<Uint8Array>);
        set_trace_id(trace: number): void;
        trace_id(): number | undefined;
        convert_to<RBC extends BodyContent>(map: (t: BC) => BuckyResult<RBC>): BuckyResult<ObjectMutBody<DC, RBC>>;
        prev_version(): Option<HashValue>;
        update_time(): JSBI;
        content(): BC;
        user_data(): Option<Uint8Array>;
        set_update_time(value: JSBI): void;
        increase_update_time(value: JSBI): void;
        set_userdata(user_data: Uint8Array): void;
        raw_measure(ctx: NamedObjectBodyContext, purpose?: RawEncodePurpose): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx: NamedObjectBodyContext, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        encode_to_buf(purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
        raw_hash_encode(): BuckyResult<HashValue>;
}
// ...
export class ObjectSigns implements RawEncode {
        constructor(desc_signs: Option<Vec<Signature>>, body_signs: Option<Vec<Signature>>);
        desc_signs(): Option<Signature[]>;
        body_signs(): Option<Signature[]>;
        reset_desc_sign(sign: Signature): void;
        reset_body_sign(sign: Signature): void;
        push_desc_sign(sign: Signature): void;
        push_body_sign(sign: Signature): void;
        latest_desc_sign_time(): JSBI;
        latest_body_sign_time(): JSBI;
        raw_measure(ctx: NamedObjectContext): BuckyResult<number>;
        raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
        static default(): ObjectSigns;
}
/**
    * NamedObject 的签名解码器
    */
export class ObjectSignsDecoder implements RawDecode<ObjectSigns> {
        raw_decode(buf: Uint8Array, ctx: NamedObjectContext): BuckyResult<[ObjectSigns, Uint8Array]>;
}
/**
    * 强类型命名对象Id
    */
export class NamedObjectId<DC extends DescContent, BC extends BodyContent> implements RawEncode, Compareable<NamedObjectId<DC, BC>> {
        constructor(obj_type: number, object_id: ObjectId);
        get object_id(): ObjectId;
        hashCode(): symbol;
        equals<ODC extends DescContent, OBC extends BodyContent>(other: NamedObjectId<ODC, OBC>): boolean;
        into(): ObjectId;
        gen(object_id: ObjectId): void;
        toString(): string;
        toJSON(): string;
        to_string(): string;
        to_base_58(): string;
        raw_measure(ctx?: any): BuckyResult<number>;
        raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
// ...
export class NamedObjectDecoder<DC extends DescContent, BC extends BodyContent, O extends NamedObject<DC, BC>> implements RawDecode<NamedObject<DC, BC>> {
        constructor(desc_content_decoer: DescContentDecoder<DC>, body_content_decoder: BodyContentDecoder<BC>, obj_builder: new (...constructorArgs: any[]) => O);
        raw_decode(buf: Uint8Array): BuckyResult<[O, Uint8Array]>;
        from_raw(buf: Uint8Array): BuckyResult<O>;
        from_hex(hex: string): BuckyResult<O>;
}
// ...
export enum ObjectCategory {
    Standard = "standard",
    Core = "core",
    DecApp = "dec_app"
}
// ...
export type ObjectIdInfo = (StandardObjectIdInfo | CoreObjectIdInfo | DecAppObjectIdInfo) & ObjectIdInfoMatcher;
export class ObjectId implements RawEncode, Compareable<ObjectId> {
    m_buf: Uint8Array;
    m_base58?: string;
    constructor(buf: Uint8Array);
    get object_id(): void;
    static default(): ObjectId;
    static copy_from_slice(buf: Uint8Array): ObjectId;
    obj_type_code(): ObjectTypeCode;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    as_slice(): Uint8Array;
    clone(): ObjectId;
    length(): number;
    toString(): string;
    toJSON(): string;
    to_string(): string;
    to_base_58(): string;
    hashCode(): symbol;
    equals(other: ObjectId): boolean;
    eq(other: ObjectId): boolean;
    static from_str(s: string): BuckyResult<ObjectId>;
    static from_base_58(s: string): BuckyResult<ObjectId>;
    to_hash_value(): HashValue;
    object_category(): ObjectCategory;
    is_stand_object(): boolean;
    is_core_object(): boolean;
    is_dec_app_object(): boolean;
    info(): ObjectIdInfo;
}
// ...
export enum ObjectTypeCode {
    Device = 1,
    People = 2,
    SimpleGroup = 3,
    Org = 4,
    AppGroup = 5,
    UnionAccount = 6,
    Chunk = 7,
    File = 8,
    Dir = 9,
    Diff = 10,
    ProofOfService = 11,
    Tx = 12,
    Action = 13,
    ObjectMap = 14,
    Contract = 15,
    Custom = 16
}
// ...
    export type AsObject = {
        id: string,
    }
// ...
    export interface TypeMap {
        TEXT: 0;
        OBJECT: 1;
    }
// ...
export class ChunkList extends jspb.Message {
    getType(): ChunkList.TypeMap[keyof ChunkList.TypeMap];
    setType(value: ChunkList.TypeMap[keyof ChunkList.TypeMap]): void;

    clearChunkIdListList(): void;
    getChunkIdListList(): Array<Uint8Array | string>;
    getChunkIdListList_asU8(): Array<Uint8Array>;
    getChunkIdListList_asB64(): Array<string>;
    setChunkIdListList(value: Array<Uint8Array | string>): void;
    addChunkIdList(value: Uint8Array | string, index?: number): Uint8Array | string;

    getFileId(): Uint8Array | string;
    getFileId_asU8(): Uint8Array;
    getFileId_asB64(): string;
    setFileId(value: Uint8Array | string): void;

    hasHashMethod(): boolean;
    clearHashMethod(): void;
    getHashMethod(): ChunkList.HashMethodMap[keyof ChunkList.HashMethodMap];
    setHashMethod(value: ChunkList.HashMethodMap[keyof ChunkList.HashMethodMap]): void;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ChunkList.AsObject;
    static toObject(includeInstance: boolean, msg: ChunkList): ChunkList.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ChunkList, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ChunkList;
    static deserializeBinaryFromReader(message: ChunkList, reader: jspb.BinaryReader): ChunkList;
}
// ...
    export interface HashMethodMap {
        SERIAL: 0;
    }
// ...
export class FileBodyContent extends jspb.Message {
    hasChunkList(): boolean;
    clearChunkList(): void;
    getChunkList(): ChunkList | undefined;
    setChunkList(value?: ChunkList): void;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): FileBodyContent.AsObject;
    static toObject(includeInstance: boolean, msg: FileBodyContent): FileBodyContent.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: FileBodyContent, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): FileBodyContent;
    static deserializeBinaryFromReader(message: FileBodyContent, reader: jspb.BinaryReader): FileBodyContent;
}